+++ connectors/destination/__init__.py
@@ -3,10 +3,26 @@ from mcp.server.fastmcp import FastMCP
 
 def register_destination_connectors(mcp: FastMCP):
     """Register all destination connector tools with the MCP server."""
+    from connectors.destination.mongo import (
+        create_mongodb_destination,
+        delete_mongodb_destination,
+        update_mongodb_destination,
+    )
+
     from .astra import (
         create_astradb_destination,
         delete_astradb_destination,
         update_astradb_destination,
+    )
+    from .databricks_vdt import (
+        create_databricks_delta_table_destination,
+        delete_databricks_delta_table_destination,
+        update_databricks_delta_table_destination,
+    )
+    from .databricksvolumes import (
+        create_databricks_volumes_destination,
+        delete_databricks_volumes_destination,
+        update_databricks_volumes_destination,
     )
     from .neo4j import (
         create_neo4j_destination,
@@ -44,6 +60,21 @@ def register_destination_connectors(mcp: FastMCP):
     mcp.tool()(create_neo4j_destination)
     mcp.tool()(update_neo4j_destination)
     mcp.tool()(delete_neo4j_destination)
+
+    # Register MongoDB destination connector tools
+    mcp.tool()(create_mongodb_destination)
+    mcp.tool()(update_mongodb_destination)
+    mcp.tool()(delete_mongodb_destination)
+
+    # Register databricks destination connector tools
+    mcp.tool()(create_databricks_volumes_destination)
+    mcp.tool()(update_databricks_volumes_destination)
+    mcp.tool()(delete_databricks_volumes_destination)
+
+    # Register databricks delta table destination connector tools
+    mcp.tool()(create_databricks_delta_table_destination)
+    mcp.tool()(update_databricks_delta_table_destination)
+    mcp.tool()(delete_databricks_delta_table_destination)
 
     # Register Pinecone destination connector tools